ref: move src/types/utils to shared/types/utils
